Move FIM token configurations from Vim config to separate config files
This makes it easiert to use various models. The user should not need to
configure this, the plugin simply gets shipped with matching config
files for the most popular models.

The plugin uses a simple '<FILL_IN_HERE>' tag which gets replaced with
the correct FIM tokens in the python code.