add g:ollama_venv_site_packages options for installation on macOS
This commit adds a new option `g:ollama_venv_site_packages` to allow
users to specify the path of the site-packages directory within the
virtual environment, which is useful for macOS where system wide pip
packages are not suggested and venv virtual environment is recommended
for isolated environments. Meanwhile there is no good way for vim to
respect the venv vironmens natively so this option is provided to
mitigate the issue for macOS users.