VimHelper: ShowTextAbove(): Use single quotes string
Doubled quotes strings interpret backslash escapes.  As currently
written, it means that if a string being modified contains a backslash,
the plugin will fail to add it as text property.

One approach would be to do more escaping, but single quoted strings are
specifically designed for text that needs to be interpreted literally.