doc: No numbered lists, tags for commands and variables
Numbered lists for individual commands and variables need to be updated
manually each time a command or a variable is added or removed.  Which
creates a noticeable maintenance when changes are published in parallel.

Users are unlikely to use these numbers, as it is much more convenient
to reference commands and global variables by their names.  Plus order
in these lists is quite subjective.

Vim documentation only numbers top level sections.

At the same time, it does make sense to add tags to allow users to
quickly find commands and global variables via the `:help` completion.